backend/requirements.txt
# AI Financial Copilot — backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
yfinance>=0.2.40
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
